requirements.txt
streamlit
pandas
numpy
scikit-learn
spacy
sentence-transformers
langchain
langchain-openai
python-dotenv
